services/resume-service/requirements.txt
gliner==0.2.22
google-generativeai==0.8.5
python-docx==1.2.0
python-dotenv==1.2.1
pdfminer.six==20250506
